feat(gossip.proto): add JoinRequest, JoinResponse, LeaveRequest, LeaveResponse message types
chore(gossip.pb.go): update generated Go code for Gossip service
feat(gossip.pb.go): add new message types for JoinRequest, JoinResponse, LeaveRequest, LeaveResponse, and LeaveResponse
feat(gossip.pb.go): add support for self field in JoinRequest, LeaveRequest, and LeaveRequest
feat(gossip.pb.go): add support for LeaveRequest message type
feat(gossip.pb.go): add support for LeaveResponse message type
feat(gossip.pb.go): add support for LeaveResponse message type
feat(gossip.pb.go): add support for LeaveResponse message type
feat(gossip.pb.go): add support for LeaveResponse message type
feat(gossip.pb.go): add support for GossipService Ping method
feat(gossip.pb.go): add support for PingRequest and PingResponse message types
feat(gossip.pb.go): add support for Join method
feat(gossip.pb.go): add support for JoinRequest and JoinResponse message types

chore(gossip.pb.go): update number of message types to 14 in gossip.proto
feat(gossip.pb.go): add JoinRequest, JoinResponse, LeaveRequest, and LeaveResponse
message types to gossip.proto

chore(gossip): add Join and Leave procedures to GossipService
feat(cache): change ReportCacheHealth to Record in memory cache
feat(cache): change ReportCacheHit to Record in metrics middleware
refactor(ratelimit): remove logging of latency in ratelimit functions
refactor(service): update comments in Listen method regarding timeouts

feat(gossip): implement cluster package for managing gossip communication

feat(gossip): add cluster struct with methods for managing cluster members and
communication

feat(gossip): implement Config struct with default values for gossip
configuration

feat(gossip): add New function to create a new cluster instance with provided
config

feat(gossip): implement run method to start the cluster's gossip loop and
background tasks

feat(gossip): add methods for joining, shutting down, and subscribing to
cluster events

feat(gossip): implement randomPeers method to select random peers for gossip
communication

feat(gossip): add methods to add and remove members from the cluster state

feat(gossip): implement gossip method for exchanging membership information
with peers

feat(gossip): add clusterServer struct to handle gossip service
feat(gossip): implement CreateHandler method in clusterServer
feat(gossip): implement Serve method in clusterServer
feat(gossip): implement Join method in clusterServer
feat(gossip): implement Leave method in clusterServer
feat(gossip): implement Ping method in clusterServer
feat(gossip): implement IndirectPing method in clusterServer
feat(gossip): implement SyncMembers method in clusterServer
feat(gossip): add Member struct with Hash method
feat(gossip): add Cluster interface with SubscribeJoinEvents and SubscribeLeaveEvents
feat(gossip): add tests for cluster membership and event emission

feat(gossip): add rpc.go file with functions for handling join, leave, ping,
indirectPing, and syncMembers in the gossip package

feat(gossip): add test for membership changes propagation to hash ring
feat(gossip): add test utility function to simulate failure in cluster
fix(load): change variable name from SystemLoadReport to SystemLoad
fix(metrics): refactor axiom methods to use Record instead of specific reports
feat(metrics): add Metric interface for all metrics to implement
feat(proto): add Join and Leave RPC methods to GossipService
feat(scripts): add deploy script for scaling regions in fly platform

refactor(ratelimit): remove unused time tracking code in service methods
refactor(ratelimit): update ratelimit fields to use integer type instead of number
feat(api): update key schema fields to use integer type instead of number
feat(api): update ratelimit fields to use integer type instead of number

fix(routes): specify that limit and duration fields should be integers

+++ apps/api/src/routes/v1_apis_deleteKeys.ts
@@ -39,7 +39,7 @@ const route = createRoute({
       content: {
         "application/json": {
           schema: z.object({
-            deletedKeys: z.number().openapi({
+            deletedKeys: z.number().int().openapi({
               description: "The number of keys that were deleted",
             }),
           }),
